tty: serial: imx: disable DCD and RI interrupts
This is a squash of cherry-picks of 4151518afa8b & 02439005d0dc
The later moves all the first did to a different function and
completes it.

4151518afa8b:

If the UART is in DTE mode, the signals DCD and RI are inputs. In
this case, the control bits in UCR3_DCD and UCR3_RI control the
interrupt of those two inputs. The two bits are 1 on reset, hence
leading to an interrupt if one of those signal changes... However,
as of now the interrupt handler does not handle these interrupts,
leading to a interrupt strom.

Solve the issue by explicitly disabling the two interrupts during
initialization.

Signed-off-by: Stefan Agner <stefan.agner@toradex.com>
Acked-by: Max Krummenacher <max.krummenacher@toradex.com>

02439005d0dc:
tty: serial: imx: fix merge inconsistency

Missing commit 688f897ddd96 ("Merge remote-tracking branch
'linux-fslc/4.9-1.0.x-imx' into toradex_4.9-1.0.x-imx-next").

Signed-off-by: Marcel Ziswiler <marcel.ziswiler@toradex.com>
Acked-by: Max Krummenacher <max.krummenacher@toradex.com>

Conflicts:
	drivers/tty/serial/imx.c
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>