can: mcp25xxfd: Add Can transmission support
Add un-optimized Can2.0 and CanFD transmission support.

On a Rpi3 we can saturate the CAN bus at 1MHz transmitting Can2.0
frames with DLC=0 for the number of configured tx-fifos.
Afterwards we need some time to recover the state before we can
fill in the fifos again.

With 7 tx fifos we can: send those 7 frames in 0.33ms and then we
wait for 0.26ms so that is a 56% duty cycle.

With 24 tx fifos this changes to: 1.19ms for 24 frames and then we
wait for 0.52ms so that is a 70% duty cycle.

--

Changelog:
      V6 -> V7: added can transmission into a separate patch to reduce
                  the size of the individual patches to make the linux-can
                  mailing-list happy...
      V7 -> V8: reduce default number of tx fifos to 6 to avoid HW-bug
                  during spi-read of TXREQ
                include feedback by Joe Burmeister
                sparse fixes based on feedback by Marc Kleine-Budde

Signed-off-by: Martin Sperl <kernel@martin.sperl.org>
(cherry picked from commit 69a74571ed18c82f0c5d8cc8e8ce5d6214070481)
(cherry picked from commit c428a5b90cddba52796f5a266c4f4e7eca71903d)