arm: dts: imx6q apalis: ov5640: make sure the mclk is on
When the parallel camera kernel modules are loaded mclk is switched on
by them, so this bug has been masked.

While at it drop the pwms property which was what clock 147 was pointing
to. Setting pwms results in the pwm being configured with 50% duty cycle
and then no longer touched by the driver.
In our HW design we use CCM_CLKO2 as the MCLK, not PWM2.

Related-to: ELB-2899
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>