regulator: pca9450: disable buck123_dvs
The current driver does not handle/use the combined BUCK123_DVS register.
So do not use the voltage setpoints for buck1, buck2 and buck3 from
that register but rather take the values from the per buck converter
registers PCA9450_BUCKxOUT_DVSy.

Fixes buck1, buck2 and buck3 to be at their reset defaults of 850 mV
independent of what is commanded from software if the PMIC is not yet
configured accordingly in the bootloader.

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>