regulator: pfuze100: fix system hang on suspend
The shutdown() callback of i2c client device calls for both scenarios,
power off and reboot the system.
In case of rebooting, the PMIC is configured to power off the power rails
when receiveing the PMIC_STBY_REQ request.
It leads to powering the module down and system hanging while suspending
the system after soft reboot.
Remove the shutdown() callback. Moreover, it is not needed now as the
power-down preparation of PMIC is done in the pfuze_power_off_prepare()
function.

Related-to: ELB-3473
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>