arm64: dts: colibri-imx8x: add lcdif display parts
Return back the definitions of parts of the lcdif display and all
related stuff to the main devicetree to keep an overlay as simple
as possible. As a panel is not a part of boards, move the panel_dpi
stuff to the main devictree.
Remains it disabled in the main devicetree.

Related-to: ELB-3246
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>