net: fec: make sure the phy hw gets re-initalised on resume
The PHY provides configurable skew/delay for RGMII. Make sure that on
resume if the PHY's power supply was cut the PHY's registers are
initialized again.
While at it move the related phy irq disable right before setting
the pins to sleep state symmetrical to how the phy irq enable is
placed in resume.

Related-to: ELB-3340
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>