arm64: config: toradex: set delay for thermal emergency poweroff
If the delay is not set, the thermal emergency poweroff function
disabled. It leads to unable powering down the system in case of
critical temperature is reached and system tool /sbin/poweroff
is unavailable or fail.
Set the option CONFIG_THERMAL_EMERGENCY_POWEROFF_DELAY_MS to
10 seconds. Usually it's enough to shutting system down with
/sbin/poweroff.
The forceful kernel-side emergency poweroff procedure performs only
if the system tool fails to power down the system within 10 seconds.

Related-to: ELB-3852
Signed-off-by: Oleksandr Suvorov <oleksandr.suvorov@toradex.com>