arm: dts: colibri-imx7: adjust reserved memory for m4
Aligning the carved out memory to ARM Linux' PMD (page middle
directory) size of 2MiB and don't have the end of the section
at the same address as the end of the DDR RAM.

Otherwise Linux truncates the used memory with the following output
on a Colibri iMX7D 512MB.

[    0.000000] Ignoring RAM at 0x8fe00000-0xa0000000
[    0.000000] Consider using a HIGHMEM enabled kernel.

and, if relocating the device tree blob to the end of the memory,
Linux will not boot without any output as it cannot access the dtb.

This is caused by the backport of 'commit fb326c6ce0dc ("fdt: Properly
handle "no-map" field in the memory region")' in kernel 'v5.4.102'.

Compare with the former solution in U-Boot:
http://git.toradex.com/cgit/u-boot-toradex.git/commit/board/toradex/colibri_imx7/colibri_imx7.c?h=2016.11-toradex&id=23b5f7c09292256f1b40572a24451778d918330d

Related-to: ELB-3958
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>