colibri_imx7_defconfig: reduce cma size
With the reserved regions for M4 and rpmsg use there is not enough
continuous memory left to reserve 256 MiB.

For modules with 256MB RAM the 50% of available memory rule limits the
size to 128MB, i.e. this commit does not further reduce the CMA size.
For modules with 1GB of RAM this also reduces the CMA size to 192MB
even though 256MB would be possible.

Prevents getting the following error message:
[    0.000000] cma: Failed to reserve 256 MiB

Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>