imx8mp: native hdmi: allow pixelclocks being slightly off
The HDMI PHY driver currently can only generate a discrete set of
pixelclock frequencies. Any mode read from a monitor's EDID with a non
matching frequency is rejected.

Change that to find the closest possible frequency and accept the mode
if it deviates less than 6%. 6% is chosen as with that any pixelclock
in the range of 22.25Mhz - 297MHz should be accepted.

Related-to: ELB-4015
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>