arm64: dts: verdin-imx8mp: rework usb vbus enable
The USB VBus enable signals can be handled by the HW dedicated VBUS
functionality and its driver. There is no need to modulate it with a
regulator with GPIO.

For the USB1 in peripheral mode the current solution does not switch off
VBUS which leads to backfeeding during shutdown.

Related to: ELB-4002
Signed-off-by: Max Krummenacher <max.krummenacher@toradex.com>