arm64: config: apalis-imx8: build ata drivers as modules
It seems like NXP iMX8 has an issue with PCIe device initialization.
In the case of multiple devices connected to the PCIe bus (that is a
common case) and all of them using the same clock source (even if
clocks are different) it seldomly leads to the inability to init the
device due to a problem with the clock restoring after the previous
device initialization. The problem is workaround by moving the SATA
device initialization to a later time i.e. by compiling it as a kernel
module (CONFIG_SATA_AHCI=m and CONFIG_AHCI_IMX=m).

Related-to: ELB-3431

Signed-off-by: Oleksii Bidnichenko <oleksii.bidnichenko@toradex.com>