ARM64: dts: apalis-imx8: Add fsl,imx8qm compatible to ixora v1.2 dts
With long-term tests we saw, that without this compatible string the
module sometimes crashes.

Searching through the code we found out that in
drivers/soc/imx/soc-imx-scu.c there is a workaround activated with that
that compatible (TKT340553_SW_WORKAROUND).

This commit adds the fsl,imx8qm compatible to the QuadPlus variant of a
device tree which obviously is needed due to that workaround.

Related-to: ELB-4112
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>