arm64: dts: apalis-imx8: leave phy-reset in push-pull
In order to prevent backfeeding we have a sleep-state of our pins.
Currently the phy-reset is put to open-drain input, High-Z and
pull-down. This causes a pull-up on hardware and the configured
pull-down to work against each other and causing some strange voltage
level.

Leave the phy-reset signal configured as push-pull and pull-up
as the reset also needs to be high during runtime suspend.

Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>