arm64: dts: apalis-imx8: remove fec regulator
Despite this being present on the hardware, remove it from the
device tree.
If this is present then the fec does power-off the regulator when going
into suspend. This leads to a non working phy after resume due to the
reset line that is not triggered. We try to fix this upstream but this
seems to become a bigger task so introduce this intermediate solution
of just removing the regulator.

According to the KSZ9131 datasheet, the power-consumption in software
power-down is about 26mW.

Related-to: ELB-4251
Signed-off-by: Philippe Schenker <philippe.schenker@toradex.com>