can: mcp25xxfd: Fix failure to probe driver when bus occupation is high
The mcp2518fd chip needs to detect a bus idle condition before changing
modes. In conditions where the bus load is high (> 50%), the mode switch
can fail. If this happens during initialization, the driver fails to
probe.

The problem is that the default bit timing configurations are done for
500kbps bus rate with a SYSCLK of 40MHz. If the bus that the device is
connected to is configured at a higher bitrate and is loaded, it can be
that we are not able to detect a sufficiently large number of "1" bits
in order to say that the bus is idle.

This commit changes the bit timing before changing modes so that the
mcp2518fd is prepared to detect a bus idle condition even at the highest
bus rate that would be allowed.

Co-authored-by: Edward Karpicz <e.karpic@ekspla.com>
Signed-off-by: Edward Karpicz <e.karpic@ekspla.com>
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>