can: mcp25xxfd: Always reset TEF index when going out of configuration mode
This patch fixes an issue that prevents the mcp25xxfd driver from
correctly handling interface down / up cycles.
When the mcp2518fd goes into configuration mode it automatically resets
all the FIFO pointers. Reset the internal representation of the current
index by the driver when leaving configuration mode so that we don't get
failures when trying to bring up an interface that was previously
brought down.

Co-authored-by: Edward Karpicz <e.karpic@ekspla.com>
Signed-off-by: Edward Karpicz <e.karpic@ekspla.com>
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>