can: mcp25xxfd: Remove optimized handling of TEFIF interrupts
The optimized TEFIF handling functions violate the required ordering of
the instructions for correctly reading the TEF according to the datasheet:

"Before reading a TEF Object, the application must check that the TEF
is not empty by reading the CiTEFSTA register."

The datasheet also lists the following list as the correct order for
reading the TEF:
- Check that the TEF is not empty.
- Read the message from the TEF.
- Increment the TEF by setting UINC.
- Process the TEF message

The optimized functions were completely removed and the TEFIF handling
code was reorganized in order to make it clear that the correct sequence
of instructions is respected.

With the optimized functions in place, starting a load test transmitting
bursts of 4000 packets/s with a bus rate of 1Mbps would cause these errors
to show up in the log:

mcp25xxfd spi1.0 can0: tefif: fifo 2 not pending - tef data: id: 00000002 flags: 00000401, ts: 11d75a31 - this may be a problem with spi signal quality- try reducing spi-clock speed if this can get reproduced

After this happened, only a reboot would fix the issue and make the
driver work again.

With the current change in place, this issue cannot be reproduced anymore,
even when running the test for 12 hours non-stop.

Co-authored-by: Edward Karpicz <e.karpic@ekspla.com>
Signed-off-by: Edward Karpicz <e.karpic@ekspla.com>
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>