ARM: dts: imx7d-colibri: Move usdhc1-cd-slp definition to iomuxc_lpsr
The usdhc1-cd-slp-grp node in the device tree is making use of PAD_LPSR
definitions and these definitions are not compatible with the iomuxc
node. Because of that, instead of setting up GPIO1_IO00 this group is
setting up the registers for GPIO1_IO15.

Moving the group to the iomuxc_lpsr node makes the setup for the correct
pin and also makes GPIO1_IO15 available again as a standard GPIO.

Relates-to: ELB-4525
Signed-off-by: Rafael Beims <rafael.beims@toradex.com>