sc16is7xx: control TXEN gpio directly from serial driver
This adds the ability to control a TX enable GPIO line directly from the
sc16is7xx i2c serial interface driver. The SPAN panel has a GPIO connected
as a TX enable function of a EIA-485 transceiver. This feature lets us
control it directly from the driver instead via user-space which is slower.

We've added the ability to define the two tx enable pins via the DTS, and
then we control it if it's defined.

Bug: FW-1610