sc16is7xx: make workthread higher priority for txen
One of the problems with getting a fast response of the TX enable pin
is that we have to poll the tx status of the chip via a kernel work
thread. This is because we have to fetch the IRQ reason via the i2c bus,
which can't be done in an irq context.

By making the workthread higher priority, we're able to fetch the tx-done
status quicker.

Bug: FW-1610