usb-serial: control TXEN gpio directly from usb-serial driver
This adds the ability to control a TX enable GPIO line directly from the
usb serial interface driver. The SPAN panel has a GPIO connected
as a TX enable function of a EIA-485 transceiver. This feature lets us
control it directly from the driver instead via user-space which is slower.

We've added the ability to define the tx enable pins via the DTS, and then we
control those that have been defined. The control code is in the higher-level
generic code and so should be enabled for any driver that utilizes those routines.
This is the case of the ftdi_sio and the cp210x drivers which the SPAN HW2 gateway
uses.

Bug: FW-1610